Split Report Results
This component was getting pretty huge, let's split the results out into
another container.
