+++ packages/frontend/app/components/reports/course-report-results.hbs
@@ -1,0 +1,42 @@
+<div class="report-results">
+  <h2>{{t "general.results"}}</h2>
+  <div class="download">
+    <button type="button" {{on "click" (perform this.downloadReport)}} data-test-download>
+      <FaIcon @icon="download" />
+      {{t "general.downloadResults"}}
+    </button>
+  </div>
+  <table>
+    <thead>
+      <tr>
+        <th colspan="3">{{t "general.course"}}</th>
+        <th>{{t "general.date"}}</th>
+        <th colspan="2">{{t "general.sessionType"}}</th>
+        <th colspan="3">{{t "general.title"}}</th>
+        <th>{{t "general.duration"}}</th>
+      </tr>
+    </thead>
+    <tbody>
+      {{#each this.results as |activity|}}
+        <tr>
+          <td colspan="3">
+            <LinkTo @route="course" @models={{array activity.courseId}}>
+              {{activity.courseTitle}}
+            </LinkTo>
+          </td>
+
+          <td>
+            {{format-date activity.startDate month="2-digit" day="2-digit" year="numeric"}}
+          </td>
+          <td colspan="2">{{activity.sessionType}}</td>
+          <td colspan="3">
+            <LinkTo @route="session" @models={{array activity.courseId activity.sessionId}}>
+              {{activity.sessionTitle}}
+            </LinkTo>
+          </td>
+          <td>{{activity.minutes}}</td>
+        </tr>
+      {{/each}}
+    </tbody>
+  </table>
+</div>
